requirements.txt
python-telegram-bot==20.7
redis==5.0.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
psycopg2-binary==2.9.9 